# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FAVI Partner Orders API
  version: 1.0.0
  description: |
    REST API through which partner shops submit their orders and later change
    the expected delivery date.

    Partners are identified by the `partnerId` path segment. There is no
    authentication in this version (out of scope by assignment); the API stores
    the data exactly as it arrived.

    All error responses use RFC 9457 Problem Details
    (`application/problem+json`).
  contact:
    name: Daniel Hejna
servers:
  - url: http://127.0.0.1:8000
    description: Local development
tags:
  - name: Orders
    description: Partner orders

paths:
  /api/v1/partners/{partnerId}/orders:
    parameters:
      - $ref: '#/components/parameters/PartnerId'
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Submit a new order
      description: |
        Creates an order under the partner namespace. The pair
        `(partnerId, orderId)` is unique; submitting it again returns `409`
        and never overwrites the stored order.

        `totalValue` is stored as sent. It is neither recomputed from the
        product lines nor verified against them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              headphones:
                $ref: '#/components/examples/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              description: Canonical URL of the created order
              schema:
                type: string
                examples: ['/api/v1/partners/PARTNER_A/orders/ORD-2026-00001']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                headphones:
                  $ref: '#/components/examples/Order'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '409':
          description: An order with the same `partnerId` and `orderId` already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                duplicate:
                  value:
                    type: https://api.favi.test/problems/duplicate-order
                    title: Duplicate Order
                    status: 409
                    detail: Order "ORD-2026-00001" already exists for partner "PARTNER_A".
                    instance: /api/v1/partners/PARTNER_A/orders
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'

  /api/v1/partners/{partnerId}/orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/PartnerId'
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Read an order
      description: Returns the stored order. The `Location` header of `createOrder` points here.
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                headphones:
                  $ref: '#/components/examples/Order'
        '404':
          $ref: '#/components/responses/OrderNotFound'
    patch:
      tags: [Orders]
      operationId: patchOrder
      summary: Change the expected delivery date
      description: |
        Partial update following JSON Merge Patch (RFC 7396). In this version
        `expectedDeliveryDate` is the only field that can change; it is required
        in the body and must not be `null`. Unknown fields are rejected with
        `422` and a pointer to each of them.

        The request is idempotent: sending the same body twice leaves the order
        in the same state.
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/PatchOrderRequest'
            examples:
              postpone:
                $ref: '#/components/examples/PatchOrderRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PatchOrderRequest'
            examples:
              postpone:
                $ref: '#/components/examples/PatchOrderRequest'
      responses:
        '200':
          description: The updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                postponed:
                  $ref: '#/components/examples/PatchedOrder'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'

components:
  parameters:
    PartnerId:
      name: partnerId
      in: path
      required: true
      description: Partner-side identifier, any format, 1 to 64 characters
      schema:
        $ref: '#/components/schemas/ExternalId'
    OrderId:
      name: orderId
      in: path
      required: true
      description: Partner-side order identifier, any format, 1 to 64 characters
      schema:
        $ref: '#/components/schemas/ExternalId'

  schemas:
    ExternalId:
      type: string
      minLength: 1
      maxLength: 64
      examples: ['ORD-2026-00001']

    DecimalAmount:
      type: string
      description: |
        Decimal number as a string to avoid floating point loss. Up to 12
        integer digits and up to 2 fractional digits, non-negative. Responses
        always carry exactly 2 fractional digits.
      pattern: '^\d{1,12}(\.\d{1,2})?$'
      examples: ['1299.99']

    DeliveryDate:
      type: string
      format: date
      description: Expected delivery date, ISO 8601 calendar date
      examples: ['2026-06-15']

    CreateOrderProductRequest:
      type: object
      additionalProperties: false
      required: [productId, name, price, quantity]
      properties:
        productId:
          $ref: '#/components/schemas/ExternalId'
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          $ref: '#/components/schemas/DecimalAmount'
        quantity:
          type: integer
          minimum: 1

    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [orderId, expectedDeliveryDate, totalValue, products]
      properties:
        orderId:
          $ref: '#/components/schemas/ExternalId'
        expectedDeliveryDate:
          $ref: '#/components/schemas/DeliveryDate'
        totalValue:
          $ref: '#/components/schemas/DecimalAmount'
        products:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderProductRequest'

    PatchOrderRequest:
      type: object
      additionalProperties: false
      required: [expectedDeliveryDate]
      properties:
        expectedDeliveryDate:
          $ref: '#/components/schemas/DeliveryDate'

    OrderProduct:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId:
          $ref: '#/components/schemas/ExternalId'
        name:
          type: string
        price:
          $ref: '#/components/schemas/DecimalAmount'
        quantity:
          type: integer
          minimum: 1

    Order:
      type: object
      required: [partnerId, orderId, expectedDeliveryDate, totalValue, products, createdAt, updatedAt]
      properties:
        partnerId:
          $ref: '#/components/schemas/ExternalId'
        orderId:
          $ref: '#/components/schemas/ExternalId'
        expectedDeliveryDate:
          $ref: '#/components/schemas/DeliveryDate'
        totalValue:
          $ref: '#/components/schemas/DecimalAmount'
        products:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderProduct'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable identifier of the problem class
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          description: Path of the request that produced the problem

    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                type: object
                required: [pointer, message]
                properties:
                  pointer:
                    type: string
                    description: JSON Pointer (RFC 6901) to the offending field
                  message:
                    type: string

  responses:
    MalformedRequest:
      description: Body is not valid JSON
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            malformed:
              value:
                type: https://api.favi.test/problems/malformed-request
                title: Bad Request
                status: 400
                detail: Request payload contains invalid "json" data.
                instance: /api/v1/partners/PARTNER_A/orders
    UnsupportedMediaType:
      description: Request `Content-Type` is not JSON
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            xml:
              value:
                type: https://api.favi.test/problems/unsupported-media-type
                title: Unsupported Media Type
                status: 415
                detail: Unsupported format, expects "json", but "xml" given.
                instance: /api/v1/partners/PARTNER_A/orders
    ValidationFailed:
      description: Body is well-formed but violates the schema
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
          examples:
            quantity:
              value:
                type: https://api.favi.test/problems/validation-failed
                title: Validation Failed
                status: 422
                detail: One or more fields are invalid.
                instance: /api/v1/partners/PARTNER_A/orders
                errors:
                  - pointer: /products/0/quantity
                    message: This value should be greater than or equal to 1.
    OrderNotFound:
      description: No order with this `orderId` exists for this partner
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missing:
              value:
                type: https://api.favi.test/problems/order-not-found
                title: Order Not Found
                status: 404
                detail: Order "ORD-404" was not found for partner "PARTNER_A".
                instance: /api/v1/partners/PARTNER_A/orders/ORD-404

  examples:
    CreateOrderRequest:
      value:
        orderId: ORD-2026-00001
        expectedDeliveryDate: '2026-06-15'
        totalValue: '1299.99'
        products:
          - productId: SKU-001
            name: Bluetooth Headphones
            price: '129.99'
            quantity: 2
          - productId: SKU-002
            name: USB-C Cable, 2 m
            price: '9.99'
            quantity: 4
    PatchOrderRequest:
      value:
        expectedDeliveryDate: '2026-07-20'
    Order:
      value:
        partnerId: PARTNER_A
        orderId: ORD-2026-00001
        expectedDeliveryDate: '2026-06-15'
        totalValue: '1299.99'
        products:
          - productId: SKU-001
            name: Bluetooth Headphones
            price: '129.99'
            quantity: 2
          - productId: SKU-002
            name: USB-C Cable, 2 m
            price: '9.99'
            quantity: 4
        createdAt: '2026-06-01T10:15:00+00:00'
        updatedAt: '2026-06-01T10:15:00+00:00'
    PatchedOrder:
      value:
        partnerId: PARTNER_A
        orderId: ORD-2026-00001
        expectedDeliveryDate: '2026-07-20'
        totalValue: '1299.99'
        products:
          - productId: SKU-001
            name: Bluetooth Headphones
            price: '129.99'
            quantity: 2
          - productId: SKU-002
            name: USB-C Cable, 2 m
            price: '9.99'
            quantity: 4
        createdAt: '2026-06-01T10:15:00+00:00'
        updatedAt: '2026-06-02T08:00:00+00:00'